feat(main): exposed private API via main.js
Added generateImages function to make use of the lib functionality via node scripting

fix #5
